Revert "Refactor the prediction into the prediction mode"
This reverts commit c30e88a4a0970d8a361515d899b1b902df36261f.
